ped version

+++ app/src/main/java/kr/ac/korea/oku/emergency/data/local/model/Destination.kt
@@ -35,5 +35,5 @@ data class Dest(
 
     val lon : Double,
 
-    val distance : Int
+    val distance : Double
 )
